✅ Task 7.2完了: StatisticsPerformance.test.js分割 - Main Controller Pattern適用
主要成果:
- ファイルサイズ削減: 3,156語 → 524語（83%削減）
- 4つの専門コンポーネント作成
- MCPツール互換性: 全ファイル2,500語以下達成
- 後方互換性: 既存exports・global関数完全保持

実装したコンポーネント:
- PerformanceTestUtilities.js - パフォーマンス測定・データ生成ユーティリティ
- DataCollectionPerformanceTests.js - データ収集パフォーマンステストスイート
- AnalysisRenderingPerformanceTests.js - 分析・描画パフォーマンステストスイート
- ExportMemoryOptimizationTests.js - エクスポート・メモリ・最適化テストスイート

技術的特徴:
- Main Controller Pattern: 軽量オーケストレーター（524語）
- 依存注入: サブコンポーネントにmainTestSuite注入
- テスト実行管理: registerTests()メソッドによる統制
- エラー処理: 存在しないメソッド・コンポーネントへの安全な対応
