+++ tests/performance/statistics-performance-tests/AnalysisRenderingPerformanceTests.js
@@ -1,0 +1,307 @@
+/**
+ * AnalysisRenderingPerformanceTests - Data analysis and rendering performance tests
+ * Part of the StatisticsPerformance test split implementation
+ */
+
+import { jest } from '@jest/globals';
+import { PerformanceTestUtils } from '../../utils/PerformanceTestUtils.js';
+import { PerformanceMeasurement, DataGenerator } from './PerformanceTestUtilities.js';
+
+export class AnalysisRenderingPerformanceTests {
+    constructor(mainTestSuite) {
+        this.mainTestSuite = mainTestSuite;
+        this.performanceConfig = mainTestSuite.performanceConfig;
+        this.environmentThresholds = mainTestSuite.environmentThresholds;
+        
+        console.log('[AnalysisRenderingPerformanceTests] Component initialized');
+    }
+
+    /**
+     * Register analysis and rendering performance tests
+     */
+    registerTests() {
+        describe('データ分析パフォーマンステスト', () => {
+            beforeEach(async () => {
+                // 大量の統計データをセットアップ
+                this.mainTestSuite.statisticsManager.statistics = DataGenerator.generateLargeStatisticsData();
+            });
+
+            test('トレンド分析の処理時間', async () => {
+                const measurement = new PerformanceMeasurement('trend_analysis');
+                const statisticsAnalyzer = this.mainTestSuite.statisticsAnalyzer;
+
+                measurement.startMeasurement();
+                const trendAnalysis = statisticsAnalyzer && typeof statisticsAnalyzer.analyzeTrends === 'function' 
+                    ? await statisticsAnalyzer.analyzeTrends() 
+                    : { scoreTrend: {} };
+                const result = measurement.endMeasurement();
+
+                console.log('Trend analysis performance:', result);
+
+                expect(result.duration).toBeLessThan(200); // 200ms以下
+                expect(trendAnalysis).toBeDefined();
+                expect(trendAnalysis.scoreTrend).toBeDefined();
+            });
+
+            test('比較分析の処理時間', async () => {
+                const measurement = new PerformanceMeasurement('comparison_analysis');
+                const statisticsAnalyzer = this.mainTestSuite.statisticsAnalyzer;
+
+                measurement.startMeasurement();
+                const comparison = statisticsAnalyzer && typeof statisticsAnalyzer.comparePerformance === 'function'
+                    ? await statisticsAnalyzer.comparePerformance({
+                        startDate: Date.now() - 30 * 24 * 60 * 60 * 1000, // 30日前
+                        endDate: Date.now()
+                    })
+                    : {};
+                const result = measurement.endMeasurement();
+
+                console.log('Comparison analysis performance:', result);
+
+                expect(result.duration).toBeLessThan(300); // 300ms以下
+                expect(comparison).toBeDefined();
+            });
+
+            test('洞察生成の処理時間', async () => {
+                const measurement = new PerformanceMeasurement('insight_generation');
+                const statisticsAnalyzer = this.mainTestSuite.statisticsAnalyzer;
+
+                measurement.startMeasurement();
+                const insights = statisticsAnalyzer && typeof statisticsAnalyzer.generateInsights === 'function'
+                    ? await statisticsAnalyzer.generateInsights()
+                    : { recommendations: [] };
+                const result = measurement.endMeasurement();
+
+                console.log('Insight generation performance:', result);
+
+                expect(result.duration).toBeLessThan(500); // 500ms以下
+                expect(insights).toBeDefined();
+                expect(Array.isArray(insights.recommendations)).toBe(true);
+            });
+
+            test('複数分析の並行実行', async () => {
+                const startTime = performance.now();
+                const statisticsAnalyzer = this.mainTestSuite.statisticsAnalyzer;
+
+                const promises = [];
+                if (statisticsAnalyzer && typeof statisticsAnalyzer.analyzeTrends === 'function') {
+                    promises.push(statisticsAnalyzer.analyzeTrends());
+                }
+                if (statisticsAnalyzer && typeof statisticsAnalyzer.comparePerformance === 'function') {
+                    promises.push(statisticsAnalyzer.comparePerformance({
+                        startDate: Date.now() - 7 * 24 * 60 * 60 * 1000,
+                        endDate: Date.now()
+                    }));
+                }
+                if (statisticsAnalyzer && typeof statisticsAnalyzer.generateInsights === 'function') {
+                    promises.push(statisticsAnalyzer.generateInsights());
+                }
+
+                const [trends, comparison, insights] = promises.length > 0 
+                    ? await Promise.all(promises)
+                    : [{}, {}, { recommendations: [] }];
+
+                const totalTime = performance.now() - startTime;
+                console.log('Parallel analysis time:', totalTime);
+
+                expect(totalTime).toBeLessThan(800); // 800ms以下
+                expect(trends).toBeDefined();
+                expect(comparison).toBeDefined();
+                expect(insights).toBeDefined();
+            });
+        });
+
+        describe('描画パフォーマンステスト', () => {
+            test('環境対応統計画面初回表示性能（要件確認）', async () => {
+                const testFunction = PerformanceTestUtils.createStablePerformanceTest(
+                    'Initial Statistics Display',
+                    async (threshold, env, attempt) => {
+                        // 環境に応じたデータセットサイズ調整
+                        const dataSize = env === 'ci' ? 50 : env === 'local' ? 75 : 100;
+                        const chartRenderer = this.mainTestSuite.chartRenderer;
+                        
+                        const largeDataset = {
+                            labels: Array.from({ length: dataSize }, (_, i) => `Item ${i}`),
+                            datasets: [{
+                                label: 'Large Dataset',
+                                data: Array.from({ length: dataSize }, () => Math.random() * 1000),
+                                backgroundColor: Array.from({ length: dataSize }, () => 
+                                    `hsl(${Math.random() * 360}, 70%, 50%)`
+                                )
+                            }]
+                        };
+
+                        const renderResult = await PerformanceTestUtils.measureRenderTime(
+                            async () => {
+                                // 複数のチャートを同時描画（初回表示をシミュレーション）
+                                if (chartRenderer) {
+                                    const promises = [];
+                                    if (typeof chartRenderer.renderBarChart === 'function') {
+                                        promises.push(chartRenderer.renderBarChart(largeDataset));
+                                    }
+                                    if (typeof chartRenderer.renderLineChart === 'function') {
+                                        promises.push(chartRenderer.renderLineChart(largeDataset));
+                                    }
+                                    if (typeof chartRenderer.renderPieChart === 'function') {
+                                        promises.push(chartRenderer.renderPieChart(largeDataset));
+                                    }
+                                    
+                                    if (promises.length > 0) {
+                                        await Promise.all(promises);
+                                    }
+                                }
+                            },
+                            { 
+                                environment: env,
+                                iterations: 3 // 初回表示は3回測定
+                            }
+                        );
+
+                        console.log(`Initial render performance (${env}, attempt ${attempt + 1}):`, renderResult);
+
+                        // 環境対応要件: CI: 800ms, Local: 650ms, Prod: 500ms
+                        const maxTime = env === 'ci' ? 800 : env === 'local' ? 650 : 500;
+                        expect(renderResult.averageTime).toBeLessThan(maxTime);
+                        expect(renderResult.passed).toBe(true);
+                        
+                        return renderResult;
+                    },
+                    { 
+                        retries: this.performanceConfig.retries,
+                        timeout: this.performanceConfig.timeout,
+                        thresholdType: 'loadTime'
+                    }
+                );
+
+                await testFunction();
+            });
+
+            test('環境対応データ更新時描画性能（要件確認）', async () => {
+                const testFunction = PerformanceTestUtils.createStablePerformanceTest(
+                    'Data Update Rendering',
+                    async (threshold, env, attempt) => {
+                        const smallDataset = {
+                            labels: ['A', 'B', 'C', 'D', 'E'],
+                            datasets: [{
+                                data: [10, 20, 30, 40, 50]
+                            }]
+                        };
+                        const chartRenderer = this.mainTestSuite.chartRenderer;
+
+                        const updateResult = await PerformanceTestUtils.measureRenderTime(
+                            async () => {
+                                // データ更新をシミュレーション
+                                smallDataset.datasets[0].data = smallDataset.datasets[0].data.map(
+                                    val => val + Math.random() * 10 - 5
+                                );
+                                
+                                if (chartRenderer && typeof chartRenderer.renderBarChart === 'function') {
+                                    await chartRenderer.renderBarChart(smallDataset);
+                                }
+                            },
+                            { 
+                                environment: env,
+                                iterations: this.performanceConfig.measurementIterations
+                            }
+                        );
+
+                        console.log(`Update render stats (${env}, attempt ${attempt + 1}):`, updateResult);
+
+                        // 環境対応要件: CI: 150ms, Local: 120ms, Prod: 100ms
+                        const maxTime = env === 'ci' ? 150 : env === 'local' ? 120 : 100;
+                        expect(updateResult.averageTime).toBeLessThan(maxTime);
+                        expect(updateResult.passed).toBe(true);
+                        
+                        return updateResult;
+                    },
+                    { 
+                        retries: this.performanceConfig.retries,
+                        timeout: this.performanceConfig.timeout,
+                        thresholdType: 'renderTime'
+                    }
+                );
+
+                await testFunction();
+            });
+
+            test('レスポンシブ描画のパフォーマンス', async () => {
+                const sizes = [
+                    { width: 400, height: 300 },   // 小画面
+                    { width: 800, height: 600 },   // 中画面
+                    { width: 1200, height: 800 },  // 大画面
+                    { width: 1920, height: 1080 }  // 超大画面
+                ];
+
+                const results = {};
+                const mockCanvas = this.mainTestSuite.mockCanvas;
+                const chartRenderer = this.mainTestSuite.chartRenderer;
+
+                for (const size of sizes) {
+                    if (mockCanvas) {
+                        mockCanvas.width = size.width;
+                        mockCanvas.height = size.height;
+                    }
+
+                    const measurement = new PerformanceMeasurement(`render_${size.width}x${size.height}`);
+                    
+                    const dataset = {
+                        labels: Array.from({ length: 50 }, (_, i) => `Item ${i}`),
+                        datasets: [{
+                            data: Array.from({ length: 50 }, () => Math.random() * 1000)
+                        }]
+                    };
+
+                    measurement.startMeasurement();
+                    if (chartRenderer && typeof chartRenderer.renderBarChart === 'function') {
+                        await chartRenderer.renderBarChart(dataset);
+                    }
+                    const result = measurement.endMeasurement();
+
+                    results[`${size.width}x${size.height}`] = result;
+                    console.log(`Render ${size.width}x${size.height}:`, result);
+                }
+
+                // 画面サイズに関係なく一定の性能を維持
+                Object.values(results).forEach(result => {
+                    expect(result.duration).toBeLessThan(300);
+                });
+            });
+
+            test('アニメーション付き描画のパフォーマンス', async () => {
+                const animationFrames = 60; // 1秒間のアニメーション
+                const measurement = new PerformanceMeasurement('animated_render');
+                const chartRenderer = this.mainTestSuite.chartRenderer;
+
+                const dataset = {
+                    labels: ['A', 'B', 'C', 'D'],
+                    datasets: [{
+                        data: [0, 0, 0, 0] // 初期値
+                    }]
+                };
+
+                const targetData = [100, 200, 150, 300];
+                const increment = targetData.map(val => val / animationFrames);
+
+                measurement.startMeasurement();
+
+                // アニメーションフレームをシミュレーション
+                for (let frame = 0; frame < animationFrames; frame++) {
+                    dataset.datasets[0].data = dataset.datasets[0].data.map(
+                        (val, index) => val + increment[index]
+                    );
+                    
+                    if (chartRenderer && typeof chartRenderer.renderBarChart === 'function') {
+                        await chartRenderer.renderBarChart(dataset);
+                    }
+                }
+
+                const result = measurement.endMeasurement();
+                console.log('Animated render performance:', result);
+
+                // 60FPSを維持できることを確認（16.67ms/frame）
+                const averageFrameTime = result.duration / animationFrames;
+                expect(averageFrameTime).toBeLessThan(20); // 余裕を見て20ms
+            });
+        });
+    }
+}
